added ecimal and $

+++ database.py
@@ -34,5 +34,5 @@ class Database:
         """
         acct = self.data.get(acct_id)
         if acct:
-            return int (acct["due"]) - int (acct["paid"])
+            return float (acct["due"]) - float (acct["paid"])
         return None
